Switch nTargetTimespan at block 98400
We want to decrease nTargetTimespan, so difficulty recalculations
happen on a more regular basis. As a result nTargetTimespan & nInterval
are no longer constants, and depend on the height in the blockchain.

Whilst in theory ComputeMinWork() should handle it, setting a
checkpoint at the point nTargetTimespan changes would be very sensible.